--- pallets/governance/tests/applications.rs
@@ -1,1 +1,0 @@
-
