--- pallets/subspace/tests/registration.rs
@@ -1,631 +1,0 @@
-mod mock;
-
-use std::collections::BTreeSet;
-
-use frame_support::{assert_err, assert_noop, assert_ok};
-use mock::*;
-use sp_core::U256;
-
-use log::info;
-use pallet_subspace::{
-    Emission, Error, FloorDelegationFee, MaxAllowedModules, MaxAllowedSubnets, MaxAllowedUids,
-    MaxNameLength, MaxRegistrationsPerBlock, MinNameLength, MinStake, RegistrationsPerBlock, Stake,
-    SubnetGaps, SubnetNames, TotalSubnets, N,
-};
-use sp_runtime::{DispatchResult, Percent};
-
-/********************************************
-    subscribing::subscribe() tests
-*********************************************/
-
-#[test]
-fn test_min_stake() {
-    new_test_ext().execute_with(|| {
-        let netuid: u16 = 0;
-        let min_stake = 100_000_000;
-        let max_registrations_per_block = 10;
-        let reg_this_block: u16 = 100;
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-        let mut network: Vec<u8> = "test".as_bytes().to_vec();
-        let address: Vec<u8> = "0.0.0.0:30333".as_bytes().to_vec();
-        network.extend(netuid.to_string().as_bytes().to_vec());
-        let name = "module".as_bytes().to_vec();
-        assert_noop!(SubspaceModule::do_register(get_origin(U256::from(0)), network, name, address, 0, U256::from(0), None), Error::<Test>::NotEnoughBalanceToRegister);
-        MinStake::<Test>::set(netuid, min_stake);
-        MaxRegistrationsPerBlock::<Test>::set(max_registrations_per_block);
-        step_block(1);
-        assert_eq!(RegistrationsPerBlock::<Test>::get(), 0);
-
-        let n = U256::from(reg_this_block);
-        let keys_list: Vec<U256> = (1..n.as_u64())
-            .map(U256::from)
-            .collect();
-
-        let min_stake_to_register = MinStake::<Test>::get(netuid);
-
-        for key in keys_list {
-            let _ = register_module(netuid, key, min_stake_to_register);
-            info!("registered module with key: {key:?} and min_stake_to_register: {min_stake_to_register:?}");
-        }
-        let registrations_this_block = RegistrationsPerBlock::<Test>::get();
-        info!("registrations_this_block: {registrations_this_block:?}");
-        assert_eq!(registrations_this_block, max_registrations_per_block);
-
-        step_block(1);
-        assert_eq!(RegistrationsPerBlock::<Test>::get(), 0);
-    });
-}
-
-#[test]
-fn test_max_registration() {
-    new_test_ext().execute_with(|| {
-        let netuid: u16 = 0;
-        let rounds = 3;
-        let max_registrations_per_block = 100;
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-        assert_eq!(RegistrationsPerBlock::<Test>::get(), 0);
-        MaxRegistrationsPerBlock::<Test>::set(1000);
-        for i in 1..(max_registrations_per_block * rounds) {
-            let key = U256::from(i);
-            assert_ok!(register_module(netuid, U256::from(i), to_nano(100)));
-            let registrations_this_block = RegistrationsPerBlock::<Test>::get();
-            assert_eq!(registrations_this_block, i);
-            assert!(SubspaceModule::is_registered(netuid, &key));
-        }
-        step_block(1);
-        assert_eq!(RegistrationsPerBlock::<Test>::get(), 0);
-    });
-}
-
-#[test]
-fn test_delegate_register() {
-    new_test_ext().execute_with(|| {
-        let netuid: u16 = 0;
-        let n: u16 = 10;
-        let key: U256 = U256::from(n + 1);
-        let module_keys: Vec<U256> = (0..n).map(U256::from).collect();
-        let stake_amount: u64 = 10_000_000_000;
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-
-        SubspaceModule::add_balance_to_account(&key, stake_amount * n as u64);
-        for module_key in module_keys {
-            delegate_register_module(netuid, key, module_key, stake_amount)
-                .expect("delegate register module failed");
-            let key_balance = SubspaceModule::get_balance_u64(&key);
-            let stake_to_module = SubspaceModule::get_stake_to_module(netuid, &key, &module_key);
-            info!("key_balance: {key_balance:?}");
-            let stake_to_vector = SubspaceModule::get_stake_to_vector(netuid, &key);
-            info!("stake_to_vector: {stake_to_vector:?}");
-            assert_eq!(stake_to_module, stake_amount);
-        }
-    });
-}
-
-#[test]
-fn test_registration_ok() {
-    new_test_ext().execute_with(|| {
-        let netuid: u16 = 0;
-        let key: U256 = U256::from(1);
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-
-        // make sure there is some balance
-        add_balance(key, 2);
-        register_module(netuid, key, 1)
-            .unwrap_or_else(|_| panic!("register module failed for key {key:?}"));
-
-        // Check if module has added to the specified network(netuid)
-        assert_eq!(N::<Test>::get(netuid), 1);
-
-        // Check if the module has added to the Keys
-        let module_uid = SubspaceModule::get_uid_for_key(netuid, &key);
-        assert_eq!(SubspaceModule::get_uid_for_key(netuid, &key), 0);
-        // Check if module has added to Uids
-        let neuro_uid = SubspaceModule::get_uid_for_key(netuid, &key);
-        assert_eq!(neuro_uid, module_uid);
-
-        // Check if the balance of this hotkey account for this subnetwork == 0
-        assert_eq!(get_stake_for_uid(netuid, module_uid), 1);
-    });
-}
-
-#[test]
-fn test_many_registrations() {
-    new_test_ext().execute_with(|| {
-        let netuid = 0;
-        let stake = 10;
-        let n = 100;
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-
-        MaxRegistrationsPerBlock::<Test>::set(n);
-        for i in 0..n {
-            register_module(netuid, U256::from(i), stake).unwrap_or_else(|_| {
-                panic!("Failed to register module with key: {i:?} and stake: {stake:?}",)
-            });
-            assert_eq!(N::<Test>::get(netuid), i + 1, "Failed at i={i}",);
-        }
-    });
-}
-
-#[test]
-fn test_registration_with_stake() {
-    new_test_ext().execute_with(|| {
-        let netuid = 0;
-        let stake_vector: Vec<u64> = [100000, 1000000, 10000000].to_vec();
-
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-
-        for (i, stake) in stake_vector.iter().enumerate() {
-            let uid: u16 = i as u16;
-            let stake_value: u64 = *stake;
-
-            let key = U256::from(uid);
-            info!("key: {key:?}");
-            info!("stake: {stake_value:?}");
-            let stake_before: u64 = SubspaceModule::get_stake(netuid, &key);
-            info!("stake_before: {stake_before:?}");
-            register_module(netuid, key, stake_value).unwrap_or_else(|_| {
-                panic!("Failed to register module with key: {key:?} and stake: {stake_value:?}",)
-            });
-            info!("balance: {:?}", SubspaceModule::get_balance_u64(&key));
-            assert_eq!(get_stake_for_uid(netuid, uid), stake_value);
-        }
-    });
-}
-
-#[test]
-fn register_same_key_twice() {
-    new_test_ext().execute_with(|| {
-        let netuid = 0;
-        let stake = 10;
-        let key = U256::from(1);
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-
-        assert_ok!(register_module(netuid, key, stake));
-        assert_err!(
-            register_module(netuid, key, stake),
-            Error::<Test>::KeyAlreadyRegistered
-        );
-    });
-}
-
-fn register_custom(netuid: u16, key: U256, name: &[u8], addr: &[u8]) -> DispatchResult {
-    let network: Vec<u8> = format!("test{netuid}").as_bytes().to_vec();
-
-    let origin = get_origin(key);
-    let is_new_subnet: bool = !SubspaceModule::if_subnet_exist(netuid);
-    if is_new_subnet {
-        MaxRegistrationsPerBlock::<Test>::set(1000)
-    }
-
-    // make sure there is some balance
-    add_balance(key, 2);
-    SubspaceModule::register(origin, network, name.to_vec(), addr.to_vec(), 1, key, None)
-}
-
-fn test_validation_cases(f: impl Fn(&[u8], &[u8]) -> DispatchResult) {
-    assert_err!(f(b"", b""), Error::<Test>::InvalidModuleName);
-    assert_err!(
-        f("o".repeat(100).as_bytes(), b""),
-        Error::<Test>::ModuleNameTooLong
-    );
-    assert_err!(f(b"\xc3\x28", b""), Error::<Test>::InvalidModuleName);
-
-    assert_err!(f(b"test", b""), Error::<Test>::InvalidModuleAddress);
-    assert_err!(
-        f(b"test", "o".repeat(100).as_bytes()),
-        Error::<Test>::ModuleAddressTooLong
-    );
-    assert_err!(f(b"test", b"\xc3\x28"), Error::<Test>::InvalidModuleAddress);
-
-    assert_ok!(f(b"test", b"abc"));
-}
-
-#[test]
-fn validates_module_on_registration() {
-    new_test_ext().execute_with(|| {
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-        test_validation_cases(|name, addr| register_custom(0, 0.into(), name, addr));
-
-        assert_err!(
-            register_custom(0, 1.into(), b"test", b"0.0.0.0:1"),
-            Error::<Test>::ModuleNameAlreadyExists
-        );
-    });
-}
-
-#[test]
-fn validates_module_on_update() {
-    new_test_ext().execute_with(|| {
-        let subnet = 0;
-        let key_0: U256 = 0.into();
-        let origin_0 = get_origin(0.into());
-        // make sure that the results won´t get affected by burn
-        zero_min_burn();
-
-        assert_ok!(register_custom(subnet, key_0, b"test", b"0.0.0.0:1"));
-
-        test_validation_cases(|name, addr| {
-            SubspaceModule::update_module(
-                origin_0.clone(),
-                subnet,
-                name.to_vec(),
-                addr.to_vec(),
-                None,
-                None,
-            )
-        });
-
-        let key_1: U256 = 1.into();
-        let origin_1 = get_origin(key_1);
-        assert_ok!(register_custom(0, key_1, b"test2", b"0.0.0.0:2"));
-
-        let update_module = |name: &[u8], addr: &[u8]| {
-            SubspaceModule::update_module(
-                origin_1.clone(),
-                subnet,
-                name.to_vec(),
-                addr.to_vec(),
-                Some(Percent::from_percent(5)),
-                None,
-            )
-        };
-
-        assert_err!(
-            update_module(b"test", b""),
-            Error::<Test>::ModuleNameAlreadyExists
-        );
-        assert_ok!(update_module(b"test2", b"0.0.0.0:2"));
-        assert_ok!(update_module(b"test3", b"0.0.0.0:3"));
-
-        let params = SubspaceModule::module_params(0, &key_1);
-        assert_eq!(params.name, b"test3");
-        assert_eq!(params.address, b"0.0.0.0:3");
-
-        FloorDelegationFee::<Test>::put(Percent::from_percent(10));
-        assert_err!(
-            update_module(b"test3", b"0.0.0.0:3"),
-            Error::<Test>::InvalidMinDelegationFee
-        );
-    });
-}
-
-#[test]
-fn deregister_within_subnet_when_limit_is_reached() {
-    new_test_ext().execute_with(|| {
-        MaxAllowedModules::<Test>::set(3);
-        assert_ok!(register_module(0, 0.into(), to_nano(10_000)));
-        assert_ok!(register_module(1, 1.into(), to_nano(5_000)));
-
-        assert_eq!(Stake::<Test>::get(0, U256::from(0)), to_nano(9_996));
-        assert_eq!(Stake::<Test>::get(1, U256::from(1)), to_nano(4_996));
-
-        MaxAllowedUids::<Test>::set(0, 1);
-        MaxAllowedUids::<Test>::set(1, 1);
-
-        assert_ok!(register_module(0, 2.into(), to_nano(15_000)));
-
-        assert_eq!(Stake::<Test>::get(0, U256::from(2)), to_nano(14_996));
-        assert_eq!(Stake::<Test>::get(1, U256::from(1)), to_nano(4_996));
-
-        assert_eq!(Emission::<Test>::get(0).len(), 1);
-        assert_eq!(Emission::<Test>::get(1).len(), 1);
-    });
-}
-
-#[test]
-fn deregister_globally_when_global_limit_is_reached() {
-    new_test_ext().execute_with(|| {
-        MaxAllowedModules::<Test>::set(2);
-
-        assert_ok!(register_module(0, 0.into(), to_nano(10_000)));
-        assert_ok!(register_module(1, 1.into(), to_nano(5_000)));
-
-        assert_eq!(Stake::<Test>::get(0, U256::from(0)), to_nano(9_996));
-        assert_eq!(Stake::<Test>::get(1, U256::from(1)), to_nano(4_996));
-
-        MaxAllowedUids::<Test>::set(0, 2);
-        MaxAllowedUids::<Test>::set(1, 1);
-
-        assert_ok!(register_module(0, 2.into(), to_nano(15_000)));
-
-        assert_eq!(Stake::<Test>::get(0, U256::from(0)), to_nano(9_996));
-        assert_eq!(Stake::<Test>::get(0, U256::from(2)), to_nano(14_996));
-        assert_eq!(Stake::<Test>::get(1, U256::from(1)), 0);
-
-        assert_eq!(Emission::<Test>::get(0).len(), 2);
-        assert_eq!(Emission::<Test>::get(1).len(), 0);
-    });
-}
-
-// Names
-#[test]
-fn test_register_invalid_name() {
-    new_test_ext().execute_with(|| {
-        let network_name = b"testnet".to_vec();
-        let address = b"0x1234567890".to_vec();
-        let stake = to_nano(1);
-
-        // make registrations free
-        zero_min_burn();
-
-        // set min name lenght
-        MinNameLength::<Test>::put(2);
-
-        // Get the minimum and maximum name lengths from the configuration
-        let min_name_length = MinNameLength::<Test>::get();
-        let max_name_length = MaxNameLength::<Test>::get();
-
-        // Try registering with an empty name (invalid)
-        let empty_name = Vec::new();
-        let register_one = U256::from(0);
-        add_balance(register_one, stake + 1);
-
-        assert_noop!(
-            SubspaceModule::register(
-                get_origin(register_one),
-                network_name.clone(),
-                empty_name,
-                address.clone(),
-                stake,
-                register_one,
-                None,
-            ),
-            Error::<Test>::InvalidModuleName
-        );
-
-        // Try registering with a name that is too short (invalid)
-        let register_two = U256::from(1);
-        let short_name = b"a".to_vec();
-        add_balance(register_two, stake + 1);
-        assert_noop!(
-            SubspaceModule::register(
-                get_origin(register_two),
-                network_name.clone(),
-                short_name,
-                address.clone(),
-                stake,
-                register_two,
-                None,
-            ),
-            Error::<Test>::ModuleNameTooShort
-        );
-
-        // Try registering with a name that is exactly the minimum length (valid)
-        let register_three = U256::from(2);
-        let min_length_name = vec![b'a'; min_name_length as usize];
-        add_balance(register_three, stake + 1);
-        assert_ok!(SubspaceModule::register(
-            get_origin(register_three),
-            network_name.clone(),
-            min_length_name,
-            address.clone(),
-            stake,
-            register_three,
-            None,
-        ));
-
-        // Try registering with a name that is exactly the maximum length (valid)
-        let max_length_name = vec![b'a'; max_name_length as usize];
-        let register_four = U256::from(3);
-        add_balance(register_four, stake + 1);
-        assert_ok!(SubspaceModule::register(
-            get_origin(register_four),
-            network_name.clone(),
-            max_length_name,
-            address.clone(),
-            stake,
-            register_four,
-            None,
-        ));
-
-        // Try registering with a name that is too long (invalid)
-        let long_name = vec![b'a'; (max_name_length + 1) as usize];
-        let register_five = U256::from(4);
-        add_balance(register_five, stake + 1);
-        assert_noop!(
-            SubspaceModule::register(
-                get_origin(register_five),
-                network_name,
-                long_name,
-                address,
-                stake,
-                register_five,
-                None,
-            ),
-            Error::<Test>::ModuleNameTooLong
-        );
-    });
-}
-
-#[test]
-fn test_register_invalid_subnet_name() {
-    new_test_ext().execute_with(|| {
-        let address = b"0x1234567890".to_vec();
-        let stake = to_nano(1);
-        let module_name = b"test".to_vec();
-
-        // Make registrations free
-        zero_min_burn();
-
-        // Set min name length
-        MinNameLength::<Test>::put(2);
-
-        // Get the minimum and maximum name lengths from the configuration
-        let min_name_length = MinNameLength::<Test>::get();
-        let max_name_length = MaxNameLength::<Test>::get();
-
-        let register_one = U256::from(0);
-        let empty_name = Vec::new();
-        add_balance(register_one, stake + 1);
-        assert_noop!(
-            SubspaceModule::register(
-                get_origin(register_one),
-                empty_name,
-                module_name.clone(),
-                address.clone(),
-                stake,
-                register_one,
-                None,
-            ),
-            Error::<Test>::InvalidSubnetName
-        );
-
-        // Try registering with a name that is too short (invalid)
-        let register_two = U256::from(1);
-        let short_name = b"a".to_vec();
-        add_balance(register_two, stake + 1);
-        assert_noop!(
-            SubspaceModule::register(
-                get_origin(register_two),
-                short_name,
-                module_name.clone(),
-                address.clone(),
-                stake,
-                register_two,
-                None,
-            ),
-            Error::<Test>::SubnetNameTooShort
-        );
-
-        // Try registering with a name that is exactly the minimum length (valid)
-        let register_three = U256::from(2);
-        let min_length_name = vec![b'a'; min_name_length as usize];
-        add_balance(register_three, stake + 1);
-        assert_ok!(SubspaceModule::register(
-            get_origin(register_three),
-            min_length_name,
-            module_name.clone(),
-            address.clone(),
-            stake,
-            register_three,
-            None,
-        ));
-
-        // Try registering with a name that is exactly the maximum length (valid)
-        let max_length_name = vec![b'a'; max_name_length as usize];
-        let register_four = U256::from(3);
-        add_balance(register_four, stake + 1);
-        assert_ok!(SubspaceModule::register(
-            get_origin(register_four),
-            max_length_name,
-            module_name.clone(),
-            address.clone(),
-            stake,
-            register_four,
-            None,
-        ));
-
-        // Try registering with a name that is too long (invalid)
-        let long_name = vec![b'a'; (max_name_length + 1) as usize];
-        let register_five = U256::from(4);
-        add_balance(register_five, stake + 1);
-        assert_noop!(
-            SubspaceModule::register(
-                get_origin(register_five),
-                long_name,
-                module_name.clone(),
-                address.clone(),
-                stake,
-                register_five,
-                None,
-            ),
-            Error::<Test>::SubnetNameTooLong
-        );
-
-        // Try registering with an invalid UTF-8 name (invalid)
-        let invalid_utf8_name = vec![0xFF, 0xFF];
-        let register_six = U256::from(5);
-        add_balance(register_six, stake + 1);
-        assert_noop!(
-            SubspaceModule::register(
-                get_origin(register_six),
-                invalid_utf8_name,
-                module_name.clone(),
-                address,
-                stake,
-                register_six,
-                None,
-            ),
-            Error::<Test>::InvalidSubnetName
-        );
-    });
-}
-
-// Subnet 0 Whitelist
-
-#[test]
-fn new_subnet_reutilized_removed_netuid_if_total_is_bigger_than_removed() {
-    new_test_ext().execute_with(|| {
-        zero_min_burn();
-
-        TotalSubnets::<Test>::set(10);
-        SubnetGaps::<Test>::set(BTreeSet::from([5]));
-        assert_ok!(register_module(0, 0.into(), to_nano(1)));
-
-        let subnets: Vec<_> = N::<Test>::iter().collect();
-        assert_eq!(subnets, vec![(5, 1)]);
-        assert_eq!(SubnetGaps::<Test>::get(), BTreeSet::from([]));
-    });
-}
-
-#[test]
-fn new_subnet_does_not_reute_removed_netuid_if_total_is_smaller_than_removed() {
-    new_test_ext().execute_with(|| {
-        zero_min_burn();
-
-        TotalSubnets::<Test>::set(3);
-        SubnetGaps::<Test>::set(BTreeSet::from([7]));
-        assert_ok!(register_module(0, 0.into(), to_nano(1)));
-
-        let subnets: Vec<_> = N::<Test>::iter().collect();
-        assert_eq!(subnets, vec![(7, 1)]);
-        assert_eq!(SubnetGaps::<Test>::get(), BTreeSet::from([]));
-    });
-}
-
-#[test]
-fn new_subnets_on_removed_uids_register_modules_to_the_correct_netuids() {
-    fn assert_subnets(v: &[(u16, &str)]) {
-        let v: Vec<_> = v.iter().map(|(u, n)| (*u, n.as_bytes().to_vec())).collect();
-        let names: Vec<_> = SubnetNames::<Test>::iter().collect();
-        assert_eq!(names, v);
-    }
-
-    new_test_ext().execute_with(|| {
-        zero_min_burn();
-        MaxAllowedSubnets::<Test>::put(3);
-
-        assert_ok!(register_module(0, 0.into(), to_nano(10)));
-        assert_ok!(register_module(1, 1.into(), to_nano(5)));
-        assert_ok!(register_module(2, 2.into(), to_nano(1)));
-        assert_subnets(&[(0, "test0"), (1, "test1"), (2, "test2")]);
-
-        assert_ok!(register_module(3, 3.into(), to_nano(15)));
-        assert_subnets(&[(0, "test0"), (1, "test1"), (2, "test3")]);
-
-        assert_ok!(register_module(4, 4.into(), to_nano(20)));
-        assert_subnets(&[(0, "test0"), (1, "test4"), (2, "test3")]);
-
-        add_balance(0.into(), to_nano(50));
-        add_stake(0, 0.into(), to_nano(10));
-
-        assert_ok!(register_module(5, 5.into(), to_nano(17)));
-        assert_subnets(&[(0, "test0"), (1, "test4"), (2, "test5")]);
-
-        assert_eq!(Stake::<Test>::iter_key_prefix(0).count(), 1);
-        assert_eq!(Stake::<Test>::iter_key_prefix(1).count(), 1);
-        assert_eq!(Stake::<Test>::iter_key_prefix(2).count(), 1);
-
-        assert_eq!(N::<Test>::get(0), 1);
-        assert_eq!(N::<Test>::get(1), 1);
-        assert_eq!(N::<Test>::get(2), 1);
-    });
-}
