+++ tests/Cargo.toml
@@ -1,0 +1,56 @@
+[package]
+name = "tests"
+version = "1.0.0"
+description = "Unit testing for the Subspace runtime."
+authors = ["Commune Nucleus Team"]
+homepage = "https://commune.com"
+edition = "2021"
+license = "Unlicense"
+publish = false
+repository = "https://github.com/commune-ai/subspace"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[features]
+default = ["std"]
+std = [
+    "pallet-balances/std",
+    "frame-support/std",
+    "frame-system/std",
+    "parity-scale-codec/std",
+    "pallet-subspace/std",
+    "scale-info/std",
+]
+try-runtime = ["frame-support/try-runtime"]
+
+[dependencies]
+substrate-fixed.workspace = true
+pallet-balances.workspace = true
+sp-tracing.workspace = true
+sp-io.workspace = true
+sp-core.workspace = true
+sp-runtime.workspace = true
+sp-arithmetic.workspace = true
+bty.workspace = true
+log.workspace = true
+
+pallet-subspace = { path = "../pallets/subspace", default-features = false }
+pallet-subnet-emission = { path = "../pallets/subnet_emission", default-features = false }
+pallet-subnet-emission-api = { path = "../pallets/subnet_emission/api", default-features = false }
+pallet-governance = { path = "../pallets/governance", default-features = false }
+pallet-governance-api = { path = "../pallets/governance/api", default-features = false }
+
+frame-support.workspace = true
+frame-system.workspace = true
+
+sp-std.workspace = true
+
+parity-scale-codec = { workspace = true, default-features = false, features = [
+    "derive",
+] }
+scale-info = { workspace = true, default-features = false, features = [
+    "derive",
+] }
+
+[dev-dependencies]
